--- modules/react-native-splash-v2/meta.json
@@ -1,5 +1,0 @@
-{
-  "title": "Splash-v2 module",
-  "description": "Greet your users with an image or logo.",
-  "root": "/modules/splash-v2"
-}
