+++ modules/react-native-splash/Preview.png
