+++ modules/react-native-splash/meta.json
@@ -1,5 +1,5 @@
 {
-  "title": "Splash",
+  "title": "Splash module",
   "description": "Greet your users with an image or logo.",
   "root": "/modules/splash"
 }
